cosmetic_assessment/requirements.txt
pillow
numpy
opencv-contrib-python